libxl: move vkb device to libxl_vkb.c
Logically it is better to move vkb to
separate file as vkb device used not only by vfb
and console.
